Make rare_idx int64 dtype even when empty
When rare_class_threshold is set such that no class qualifies as rare
(train_count < threshold is False for every class), np.array([]) defaults
to float64 dtype. Subsequent indexing with rare_idx then fails:

  IndexError: arrays used as indices must be of integer (or boolean) type
    in scattering-stacking.py:188 (rec[rare_idx].mean())

This never happens in production with FIESTA-bio (95 classes,
threshold=200, several below) but surfaced in workflow smoke testing
where the synthetic fixture has 1 image per class. Forcing
dtype=np.int64 keeps an empty rare_idx well-typed; downstream
rec[rare_idx].mean() returns NaN, which is the correct semantic answer
for 'no rare classes'.

Found via fiesta-workflows bio-stacking smoke test (commit 9a9f4cb on
annefou/fiesta-workflows).

+++ tools/scattering-stacking/scattering-stacking.py
@@ -78,7 +78,8 @@ rare_idx = np.array(
         i
         for i in range(N_CLASSES)
         if train_counts.get(i, 0) < rare_class_threshold
-    ]
+    ],
+    dtype=np.int64,
 )
 print(f"Rare classes (train < {rare_class_threshold}): {len(rare_idx)}")
